refactoring for multiple pages
