+++ src/components/Gallery/CategoryName/styles.css
@@ -7,3 +7,8 @@
 
   cursor: pointer;
 }
+
+.category-name a {
+  text-decoration: none;
+  color: #000000;
+}
