+++ src/components/Gallery/styles.css
