+++ src/components/Header/index.tsx
@@ -1,3 +1,5 @@
+import './styles.css'
+
 import { Categories } from './Categories'
 import { Login } from './Login'
 import { Menu } from './Menu'
@@ -5,31 +7,18 @@ import { Place } from './Place'
 import { Search } from './Search'
 import { ShoppingCart } from './ShoppingCart'
 
-import './styles.css'
+import { CategoriesArrayProps, HandleMenu } from '../../types/types'
 
-export type CategoriesProps = {
-  category: string
-  id: number
-}
+type HeaderProps = CategoriesArrayProps & HandleMenu
 
-export type HeaderProps = {
-  categories: CategoriesProps[]
-}
-
-type Functions = {
-  handleMenu: () => void
-}
-
-export type Header = HeaderProps & Functions
-
-export function Header(props: Header) {
+export function Header(props: HeaderProps) {
   return (
     <header>
       <nav>
         <div className="top-nav">
           <div className="flex-line">
             <div className="left-top-nav">
-              <Menu onClick={props.handleMenu} />
+              <Menu handleMenu={props.handleMenu} />
               <a href="/" id="logo">
                 <img src="../../../public/logo.svg" alt="logo" />
                 E-Commerce
