--- src/components/Main/CategoryName/index.tsx
@@ -1,9 +1,0 @@
-import './styles.css'
-
-type CategoryNameProps = {
-  category: string
-}
-
-export function CategoryName(props: CategoryNameProps) {
-  return <h3 className="category-name">{props.category}</h3>
-}
