--- src/components/Main/styles.css
@@ -1,7 +1,0 @@
-main {
-  position: relative;
-  top: 18.7rem;
-
-  padding-inline: 1rem;
-  padding-bottom: 2rem;
-}
