+++ src/pages/LandingPage/styles.css
@@ -1,0 +1,7 @@
+main {
+  position: relative;
+  top: 18.7rem;
+
+  padding-inline: 1rem;
+  padding-bottom: 2rem;
+}
